requirements.txt
asgiref==3.6.0
astor==0.8.1
astropy==5.1
asttokens==2.0.8
autograd==1.6.2
autograd-gamma==0.5.0
autopep8==1.7.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
black==24.4.2
braceexpand==0.1.7
build==1.0.3
CacheControl==0.13.1
cachetools==4.2.4
certifi==2024.2.2
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorlog==6.8.2
crashtest==0.4.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.7
Django==4.1.7
django-environ==0.4.5
django-storages==1.11.1
dulwich==0.21.6
ecos==2.0.13
einops==0.8.0
executing==1.0.0
fastjsonschema==2.19.0
filelock==3.13.1
fonttools==4.37.1
formulaic==1.0.1
fsspec==2024.5.0
future==1.0.0
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==20.1.0
h5py==3.9.0
idna==3.7
imageio==2.34.1
imbalanced-learn==0.10.1
importlib-metadata==6.8.0
importlib-resources==6.1.1
installer==0.7.0
interface-meta==1.3.0
ipykernel==6.15.2
ipython==8.4.0
jaraco.classes==3.3.0
jedi==0.18.1
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==7.3.5
jupyter-core==4.11.1
keyring==24.3.0
kiwisolver==1.4.4
lazy_loader==0.4
libcst==0.4.9
lifelines==0.27.8
lightgbm==3.3.0
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
mypy-extensions==1.0.0
nest-asyncio==1.5.5
networkx==3.1
numba==0.58.1
numexpr==2.8.6
numpy==1.23.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python==4.4.0.46
osqp==0.6.7
packaging==24.0
pandas==2.0.3
paramiko==2.11.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
psutil==5.9.2
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyerfa==2.0.0.1
Pygments==2.13.0
pynndescent==0.5.12
pyodbc==4.0.34
pyparsing==3.0.9
pyproject_hooks==1.0.0
pysftp==0.2.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==23.2.1
qdldl==0.1.7.post2
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scikit-survival==0.22.2
scipy==1.10.1
seaborn==0.12.2
SecretStorage==3.3.3
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.0
sympy==1.12
threadpoolctl==3.1.0
tifffile==2023.7.10
tokenize-rt==5.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
torch==2.3.0
torchvision==0.18.0
tornado==6.2
tqdm==4.66.4
traitlets==5.3.0
triton==2.3.0
trove-classifiers==2023.11.22
typing_extensions==4.11.0
typing-inspect==0.8.0
tzdata==2024.1
umap-learn==0.5.6
urllib3==2.2.1
virtualenv==20.24.7
wcwidth==0.2.5
webdataset==0.2.86
wheel==0.43.0
wrapt==1.16.0
xgboost==1.7.5
zipp==3.17.0